decimal field prob solved

+++ finance/models.py
@@ -5,7 +5,7 @@ from erp.users.models import userprofile
 
 class Budget(models.Model):
     name = models.CharField(max_length=1)
-    total_amount = models.DecimalField(max_digits=12,decimal_places=2,blank=True,null=True)
+    total_amount = models.FloatField(blank=True,null=True)
     department = models.ForeignKey(Department)
     comment = models.TextField(blank=True)
     event_core_status = models.BooleanField(default=False)
@@ -16,12 +16,12 @@ class Item(models.Model):
     budget = models.ForeignKey(Budget)
     department = models.ForeignKey(Department)
     name = models.CharField(max_length=50)
-    quantity=models.DecimalField(max_digits=12,decimal_places=2,blank=True,null=True)
+    quantity=models.FloatField(blank=True,null=True)
     description = models.TextField(blank=True)
     comment = models.TextField(blank=True)
-    original_amount = models.DecimalField(max_digits=12,decimal_places=2,)
-    request_amount = models.DecimalField(max_digits=12,decimal_places=2,blank=True, null=True)
-    balance_amount = models.DecimalField(max_digits=12,decimal_places=2,blank=True, null=True)   
+    original_amount = models.FloatField()
+    request_amount = models.FloatField(blank=True, null=True)
+    balance_amount = models.FloatField(blank=True, null=True)   
     request_status = models.BooleanField(default=False)  
     
 class OpenBudgetPortal(models.Model):
